Add support of case-insensitive enums (#303)
Enumerations was case-sensitive only which is usually too strict in
many projects. Swagger 2.0 does not support any mechanism to
distinguish enum case-sensitivity (or more precisely, case folding).
This commit adds support to parameters and definitions. Every type may
set case-sensitivity with extension 'x-go-enum-ci'. If the options is
not set, the previous behavior is kept. Moreover, there is command line
option '--with-enum-ci' that switches all enumerations to be
case-insensitive by default (for operations only). The feature makes
sense for strings only and restricted in validate.EnumCase() function.

- Replace validation function Enum -> EnumCase
- Add 'x-go-enum-ci' vendor extension
- Add '--with-enum-ci' command line operation option
- Add test YAML specification for #303
- Add tests of extension in parameters and definitions
- Fix all tests invalidated by the change
- Update go-openapi/validate v0.19.7 -> v0.19.8

Signed-off-by: David Skorvaga <55138449+pheepi@users.noreply.github.com>